docs(editor): move tutorials into skillz directory
