+++ apps/docs/src/skillz/tutorials/getting-started/step-2/App.vue
