+++ apps/docs/src/skillz/tutorials/getting-started/step-3/App.vue
